Initial commit for Kubernetes integration with OVN.
This starts with 3 utilities and a library.

The utilities are:

1. ovn-k8s-watcher.py
This is a daemon that is run in the k8s master node (where
OVN database resides too.). This daemon listens on the
k8s API server for events. It is interested in pod creation
and deletion. It is also interested in creation and destruction
of services and endpoints to setup load balancing and gateway rules.

When a pod gets created, it creates a logical port and asks
OVN to assign a IP address and MAC address. It uses this IP address
and MAC address to annotate the pod object in the API server.

2. ovn-k8s-cni-overlay.py
This is the CNI plugin for the overlay mode and resides
in every host where containers are run. When kubelet
spawns a container, it calls the CNI plugin with a few
env variables set.  Instead of CNI plugin making call to
OVN NB to create a logical port, it calls the k8s API server
to get the pod annotation where the watcher would have set
the IP and MAC. This is designed this way to prevent direct
access to the NB database from containers.

3. ovn-k8s-overlay.py

This is a initialization script used for quick system bringup.
This script talks to OVN NB directly to create logical switches.

Signed-off-by: Gurucharan Shetty <guru@ovn.org>
Co-authored-by: Salvatore Orlando <salv.orlando@gmail.com>
Signed-off-by: Salvatore Orlando <salv.orlando@gmail.com>