Make node-name argument compulsory.
Currently, when a user does not provide a node-name argument
to the "init" scripts, we get it from the system-id of Open_vSwitch
database.  We have a requirement that the same thing needs to be passed
to kubelet too.  It is easy for a user to miss this if the argument is
optional.

Signed-off-by: Gurucharan Shetty <guru@ovn.org>