Avoid continuosly handling events due to leader election annotations
Leader election causes a lot of endpoint updates (see kubernetes
issue #23812). This implies that roughly every 2 seconds an endpoint
MODIFIED event is sent for processing, for every endpoint where
leader election is enabled (examples: kube-scheduler or
kube-controller-manager).

This patch leverages the so far unused object cache in the endpoint
watcher, sending the event only when ip changes are detected in the
endpoint data structure. To this aim, the logic for extracting ips
from the endpoint object has been moved from overlay.py into
endpoint_watcher.py

Also, tt is not to be taken for granted that given and endpoint,
there will always be a service with that name. This patch therefore
also adds explicit handling of 404 errors.
This is however a rather minor point as the final result of this
error handling is the interruption of endpoint processing in any
case.

Signed-off-by: Salvatore Orlando <salv.orlando@gmail.com>