Add flake8 tox environment.
flake8 can be used to check coding style and some basic error
checking.  tox is used to run flake8 in a Python virtual environment
and may prove useful for setting up other test environments in the
future.  It helps ensure we all run with the same versions of
dependencies.

The patch also includes a few trivial changes to make flake8 run
cleanly.

Running flake8 can be done with:

    $ tox -e flake8

Signed-off-by: Russell Bryant <russell@ovn.org>