ovn-k8s-overlay: Flush the IP address of physical interface.
When we add a physical interface for a gateway, we copy
its IP address and MAC address and make it the IP address
and MAC address of the gateway router port in the logical space.

In the physical space, we add that interface as a port of br-int.
Once we do that, we should flush its IP address.

Signed-off-by: Gurucharan Shetty <guru@ovn.org>