Do not assume pod is always in cache on delete
Replace del statement with pop method, in order to avoid watcher
failures due to KeyError when deleting a pod.
There is at least one corner case where a pod is never processed
and therefore never stored in cache, and that would be when the
kube scheduler is down and does not come up before the pod is deleted.

It is worth noting however that the observed failure is not
critical as the watcher is restarted.

Signed-off-by: Salvatore Orlando <salv.orlando@gmail.com>