ovn_k8s: Sync k8s objects with OVN objects.
When the watcher dies, it is possible that some
objects in k8s get deleted. When the watcher starts
again, we will miss those events and stale entries
will remain in the OVN database.

With this commit, we do the following steps.
1. Make a connection to k8s API server to get any
events from that time on. But we do not process
any of the events till step 3.

2. We connect to k8s API server and get all the
pod and services objects. We then compare it
with OVN database to sync them.

3. We now start processing new events from K8s API
server.

Signed-off-by: Gurucharan Shetty <guru@ovn.org>