Use network namespace from CNI arguments
Use the network namespace path provided via CNI arguments
thus removing the need for retrieving the container pid
and then building the namespace path.

The network namespace in /var/run/netns will list container ids
rather than pids. This could cause issues on the DELETE action
for containers created before this patch was applied.

It is recommended to update the plugin on nodes when no container
is running on that node.

Signed-off-by: Salvatore Orlando <salv.orlando@gmail.com>