CNI plugin: Adopt pyroute2
Configure the container interface using the pyroute2 library.
With this change, the CNI plugin now only needs to shell out for
creating a symlink to the network namespace (unfortunately pyroute2
assumes a namespace always has a reference in /var/run/netns).

The patch uses IPDB. It could have implemented also using IPRoute
and NetNS. IPDB was chosen simply because it has a syntax which
is easier to read. Switch to IPRoute/NetNS is however possible at
any time.

This patch also reorganizes a bit setup_interface in two logical
steps: in the first one the veth pair is created and one interface
move into the network namespace; the second step only deals with
configuration of the container interface.

Signed-off-by: Salvatore Orlando <salv.orlando@gmail.com>