create_logical_port: return if pod name not specified
Code will try to create a logical port even if the pod name
is not found in pod data. While this is however extremely
unlikely to occurr, if not impossible, this patch simply ensures
the if condition actually tests what we want to test.

The logical_port variable is indeed built from namespace and pod
name and will never be false in a test.

Signed-off-by: Salvatore Orlando <salv.orlando@gmail.com>