delete_logical_port: abort if pod name not found
The current code uses a string built from namespace and pod
name for verifying whether it's able to delete a logical port.

This string however can never fail a test. The test should check
the pod name. The condition in which a pod name is however empty
or None is extremely unlikely if not impossible.

This patch also contains minor improvements in log statements
in the same routine.

Signed-off-by: Salvatore Orlando <salv.orlando@gmail.com>