ovn-k8s-overlay: Make management port and routes persistent.
Currently, we create the management port on each node and assign
it IP addresses and then create routes. If there is a reboot,
one has to manually re-run the *-init scripts (or add it to custom
startup scripts) to re-add the IP addresses and routes.

With this commit, we identify whether the system is either Debian
based or redhat based and add it to the network configuration
scripts of the platforms.

Signed-off-by: Gurucharan Shetty <guru@ovn.org>