ovn-k8s-overlay: Configure NB database via SSL too.
Currently ovn-k8s-overlay connects to OVN NB database
over TCP connections for initial configuration. One could
potentially then close the TCP connection as we don't need
it afterwards.  We have the ability to configure the same
using SSL too, so as well add that ability.  This patch
does the same.

Signed-off-by: Gurucharan Shetty <guru@ovn.org>