ovn-k8s-overlay: Add Windows minion init support
Currently there is no Windows support when running the
minion init script.

This patch adds the posibility to run minion init on Windows
nodes in order to add them as minion nodes to a kubernetes
deployment.

The logic is the same as in Linux case, the difference is that
the Linux specific commands are replaced with their equivalent
powershell commands on Windows.

Signed-off-by: Alin-Gheorghe Balutoiu <abalutoiu@cloudbasesolutions.com>