Fix typo inside sync_pods function
There is a typo inside the sync_pods function.

When the logical port is created, it will have "external-ids:pod=true",
but inside the sync_pods function it will try to fetch existing logical
ports based on "external_id:pod=true" which will always return nothing
since the logical ports are not created with this tag.

The fix consists in changing the query to retrieve all the logical
ports having "external-ids:pod=true" instead of "external_id:pod=true".

Signed-off-by: Alin Balutoiu <abalutoiu@cloudbasesolutions.com>