ovn-k8s-overlay: Specify the outport for the external gateway.
Specifying the outport is not mandatory as OVN can figure this
out based on the IP address of the gateway.  But there are
environments like GCE where the default gateway is not in
the same subnet as the VM's IP address. In cases like that,
specifying the outport is useful.

Signed-off-by: Gurucharan Shetty <guru@ovn.org>