config: Cache the config values and fix Windows support
Relying on pip to place the config in a specific path does
not seem to be reliable.

We will now expect to find the config file in a default path.
The default path is "/etc/openvswitch/ovn_k8s.conf" on Linux
and "C:\etc\ovn_k8s.conf" on Windows.
If the config file is not found there, we are using the default
values which are stored in a dictionary inside config.py.

Signed-off-by: Alin-Gheorghe Balutoiu <abalutoiu@cloudbasesolutions.com>