vendor: Rebase to Kubernetes 1.7.
It looked very complex to indivdually fetch and update
packages using govendor to make the vendor directory compliant
with k8s 1.7.  The main complications are around how govendor
does a fetch from master branch of some dependent repositories.
Also, k8s apimachinary, client-go, apis etc look to have changes
where the versionining is quite poor. For e.g., it is hard to
know which version of apimachinary works with which version of
client-go repository.

It looks much easier to just copy the vendor directory from
a particular released kubernetes version.  In addition to
kubernetes, we have very less external dependencies.

This commit is a squash of the following steps:

1. Cleanup current vendor directory with a 'rm -rf'.
2. Copy k8s.io/vendor contents recursively to
our vendor directory.  This does get some code
which we don't need, but hopefully that is not a big deal
3. govendor init
4. govendor fetch github.com/openshift/origin/pkg/util/netutils
5. govendor fetch github.com/urfave/cli

This is enough to compile.

Signed-off-by: Gurucharan Shetty <guru@ovn.org>