vagrant: Add extra wait times for k8s daemons.
With the changeover to etcd 3.x cluster, it has been
observed that k8s-apiserver takes a longer time to
initialize. If it does not initialize in time, other
daemons that connect to it fail. There is clearly a
better way to script this up, but for the time being
increase the timeout.

Signed-off-by: Gurucharan Shetty <guru@ovn.org>