go-controller: Change node_name to be always lowercase
When containers are created through kubernetes, an event will be sent.
The event will contain the lowercase hostname even though kubernetes
even though the hostname contains also some uppercase letters.

This causes the watcher to fail when it tries to fetch the logical switch
based on the hostname from kubernetes. It will fail to create the logical
port for the pod. Switching to lowercase hostname when creating the
logical switch will fix the issue.

Signed-off-by: Alin Balutoiu <abalutoiu@cloudbasesolutions.com>