gateway-init: Use a lock to prevent address reuse.
The previous assumption was that the gateway-init
would be called manually. So there was no chance of
race conditions when addresses in 100.64.1.0/24 subnet
was chosen. With the attempt to run this in daemonsets,
we can't make that assumption. So provide a lock.

A different idea to consider is to annotate the kube node
object with these IP addresses too (along with switch
subnet). But that will mean that gateway-init cannot be
called independently.

Signed-off-by: Gurucharan Shetty <guru@ovn.org>