cluster/master: Fix node subnet annotation.
The current code makes an assumption that no
kubelet is running in master node. But when kubelet
does run in master, it will allocate a separate subnet
for that node and annotate it.

With this commit, we skip annotating the node
if the node is the same as the master.

Signed-off-by: Gurucharan Shetty <guru@ovn.org>