go-controller: simplify factory into just a watch factory
Instead of having it be the factory for the OVN Controller and
Cluster Controller, just make it a factory for watches that
the controllers themselves can consume.

This simplifies controller object creation somewhat since we
can now make their internal variables private and use
constructor functions in the controller packages themselves
rather than in the factory.

Signed-off-by: Dan Williams <dcbw@redhat.com>