go-controller: Add Windows support for minion-init
This patch adds support for doing minion-init on a Windows host.

The following changes were made:
- Add check for master / gateway init on Windows hosts which is
currently not supported
- Update build scripts to compile the Windows version of ovnkube.
The binary will be found in _output/go/windows/ovnkube.exe
- Moved the gatewayInit code in another file and added conditional
compilation. The gatewayInit is making use of "github.com/vishvananda/netlink"
package which does not compile on Windows.
- Implemented configureManagementPort for Windows. Added the equivalent
commands for enabling the interface, setting IP and the rest.
- Added conditional compilation for nicstobridge.go which is used
only in gatewayInit and imports the package "github.com/vishvananda/netlink"
- Added equivalent commands to start / restart the
openvswitch / ovn-controller services on Windows

Signed-off-by: Alin Balutoiu <abalutoiu@cloudbasesolutions.com>