nicstobridge: Handle the case for /32 IP address.
One could set the IP address of a Linux interface as
172.16.1.2/32 and then give it a default gateway of 172.16.1.1.
This can be done for e.g. by running the following commands.

1. ifconfig eth0 172.16.1.2 netmask 255.255.255.255 broadcast 172.16.1.2
2. route add 172.16.1.1 dev eth0
3. route add default gw 172.16.1.1

The above configuration is what google cloud uses for its VMs.

When the above configuration is used, the current code can fail
because we may attenmpt to do 3. before 2. With this commit,
we do 3. after everything else has been done.

Fixes: #224
Signed-off-by: Gurucharan Shetty <guru@ovn.org>