policy: Handle "<none>" label selectors.
With k8s 1.9, the API of newListWatchFromClient() changed
where it allows string as the labelSelector.  For this
change, we changed our code to use the function
metav1.FormatLabelSelector() which takes a metav1.LabelSelector
and returns its string equivalent. When the labelSelector is empty,
the string equivalent is "<none>". But newListWatchFromClient()
does not like when we pass the string "<none>" to it.

This commit special cases this scenario and passes an empty string
when metav1.FormatLabelSelector() returns a "<none>"

Signed-off-by: Gurucharan Shetty <guru@ovn.org>