go-controller: make config defaults clearer and fix OVN address bug
Instead of putting most of the config defaults way down
in the code, explicitly set the defaults where the global
config options are declared.  We can't do this with the OVN
options though, since the cert/key defaults depend on the
OVN database address scheme.

Clean up option overrides by using a generic function
that copies struct fields if they aren't empty, rather
than doing that explicitly on each struct field as before.

Also, a bug in the OVN auth defaults handling ("https"
vs "ssl") caused the ssl defaults to always be empty,
which exposed an issue in the testcases and requires
some testcase reorganization.

Signed-off-by: Dan Williams <dcbw@redhat.com>