config: Fix OVN address when fetching from external_ids
The OVN addresses from external_ids will have format
ssl:1.2.3.4:6641 but we need to store them in URL format,
i.e. ssl://1.2.3.4:6641.

This will cause issues when trying to read the OVN
addresses from external_ids which will cause a crash
with error "missing port in address". This happens
if the components try to read the OVN addresses from
external_ids.

This commit fixes the issue by replacing the first
occurence of ':' with '://' when fetching the address
from OVS external_ids.

Signed-off-by: Alin Balutoiu <abalutoiu@cloudbasesolutions.com>