gateway-init: Add the next hop after router port creation.
We were adding a route to the gateway router with a outport
of a gateway router port. But the gateway router was not yet
created. This resulted in some error messages in ovn-northd.log.
But were harmless as we would eventually create the port.

This commit does the route addition after the router port is
created.

Signed-off-by: Gurucharan Shetty <guru@ovn.org>