nicstobridge: Make nicstobridge transient.
When we create a gateway, we automatically add the
physical interface as a port of OVS bridge and move
the IP address from physical interface to OVS bridge.

When the machine is rebooted, this will cause OVS
to recreate the bridge - but not move the IP address
resulting in not being able to ssh to the box.

With this, we will add '--delete-transient-ports' to
/etc/default/openvswitch[-switch] so that when OVS
is restarted it will not add the physical interface
to the bridge.

You will have to run ovnkube again for it to be readded
to the bridge.

Note that, the OVS packages that we are recommending for Ubuntu
for quick-start from http://packages.wand.net.nz/ has a bug in
their systemd packaging (they don't include /etc/default/openvswitch
as part of their package). Hence the additional note in README

Signed-off-by: Gurucharan Shetty <guru@ovn.org>