go-controller: consolidate ovn-controller object watches to federated informers
Instead of creating new informers for each of the OVN policy-related watches,
use the existing informers of the given type.  This requires tracking handlers
and doing the object filtering manually in the client (rather than the server
as before) to ensure that one hander's filter does not impact other handlers
of the same type, since they are using a shared informer.

We must also manually run the ADD event when attaching a handler, which the
informer previously did for us.  But since the informer has long-since been
created and run the initial ADD by the time any handler is attached, it must
be done manually.

Signed-off-by: Dan Williams <dcbw@redhat.com>