ovnkube: race while generating gateway IPs
Let us say that we are initializing multiple gateway nodes at the same
time. The code in gateway-init.go tries to retrieve the next available gateway
IP by contacting NB. Even though there is code in place to lock NB DB for this
operation, the lock is released too soon. As a result, we end up in a
situation where all the gateway nodes get the same IP address.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>
Reviewed-by: Ram Vepa <rvepa@nvidia.com>