ovnkube: persist k8s CA certificate file path as OVS external ID
In the case where ovnkube was executed with all the information provided on
the CLI, we only persist k8s-api-server and k8s-api-token as external IDs in
Open_vSwitch table. This results in ovn-k8s-cni-overlay to fail in
InitConfig()  and as a result no networking is configured for the pod.

When SSL is in use, the CNI looks for CA certificate and it cannot find
it. This patch fixes the issue by persisting k8s-ca-certificate in
Open_vSwitch table.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>
Reviewed-by: Vit Yenukas <vity@nvidia.com>