ovn: Fix error catching when fetching the mac address
If OVS fails to retrieve the mac address of the k8s interface
created on the host, it will return "[]" instead of "".

This means that the error will not be catched and it will try
to continue which will give misleading error messages. This will
happen when it will try to add the port in OVN.

This patch also fixes some log messages which were not logging
the interface name variable.

Signed-off-by: Alin Balutoiu <abalutoiu@cloudbasesolutions.com>