ovn-kube-util: provide bridges-to-nic subcommand
This patch adds a subcommand bridges-to-nic to ovn-kube-util to restore back
the NIC interface that was added to OVS bridge as part of gateway
initialization. That is, given breth0 it moves the IP/routes associated with
breth0 to eth0 and then deletes the OVS bridge breth0.

I have renamed arguments to saveIPAddress(), delAddRoute(), and saveRoute()
so that I could reuse those functions to bridges-to-nic.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>
Reviewed-by: Vit Yenukas <vity@nvidia.com>