go-controller: Use a timeout for local unix socket calls.
Currently, we are running ovn-nbctl commands in the watcher
to program OVN NB database whenever there are events from
kubernetes. But there is no timeout for these calls. If for
some reason, the command hangs, then our watcher will be
blocked.

With this commit, we will use RunOVNNbctlUnix() function
from util package. The return type is alredy stringified
with spaces and quotes removed. So at some places, I
had to remove the string conversions and trimming spaces
and quotes.

Signed-off-by: Gurucharan Shetty <guru@ovn.org>