ovn-k8s-overlay: Ability to remove node.
You can run this on a master node:
ovn-k8s-overlay remove-node --node-name k8sminion2

Or run this on any other node:
ovn-k8s-overlay remove-node --node-name k8sminion2 \
-nb-address=$NB_ADDRESS \
-nb-client-privkey /etc/openvswitch/ovncontroller-privkey.pem \
-nb-client-cert /etc/openvswitch/ovncontroller-cert.pem \
-nb-client-cacert /etc/openvswitch/ovnnb-ca.cert

The code does not try to remove the OVS bridge on the node. The
thought process is that:
1. We already have a 'ovn-kube-util bridges-to-nic' option for that.
2. We want to be able to run this command on the master.
3. We can in the future have a node watcher in master that will
call this whenever a node gets deleted.

Signed-off-by: Gurucharan Shetty <guru@ovn.org>