contrib: Fix issues encountered with ansible playbooks
This commit fixes the issues encountered with the ansible playbooks.
It also changes the checks to fix some idempotency issues and increases
overall deployment setup time.

The errors are now handled properly and a message is displayed to
the user if something goes wrong and what it should do.

The sleeps within the playbooks waiting for a task to be complete have
also been replaced. It now checks the expected behaviour after the task.
Example: creating docker network when connection drops for a short period,
instead of sleeping for a given amount of time it waits for the WinRM
connection to come back.

Replaced the infracontainer on Windows with a nanoserver image which
should increase the speed of the deployment.

Signed-off-by: Alin Balutoiu <abalutoiu@cloudbasesolutions.com>