ovn-k8s-cni-overlay: Update Windows CNI and remove the workarounds
This commit removes the Windows specific workarounds and makes it
look more similar to the Linux code.

The endpoint name should not depend on the container details passed
to the CNI (e.g. container ID). This is because kubelet calls
at the moment the CNI for every container within a pod when it should
be called only for the sandbox container.

Signed-off-by: Alin Balutoiu <abalutoiu@cloudbasesolutions.com>