vagrant: Use a custom debian server.
We were using packages.wand.net.nz to download deb files for OVS/OVN.
But, with OVS 2.9.x, they have changes the startup scripts
to be different than what we provide in OVS. Their startup
scripts have bugs in them.

So I decided to create a new debian repo for ovn-kubernetes
project - till we have a better solution.  We need this
because OVS 2.9.1 was released last week and it has fixes
that we can take.

OVS 2.9.1 also has raft based clustering for OVSDB databases.

Signed-off-by: Gurucharan Shetty <guru@ovn.org>