gateway: Add a localnet gateway.
There are k8s deployments with only master node (as a starting
point). And sometimes they just deploy a DNS pod or some other
pods in the master node. The DNS pod in pod network will go into a
crash loop because it cannot access the k8s daemons' IP in host
network.  Even if we deploy a OVN gateway on master, this won't
help because OVN gateway will have the same IP as the host IP.

Currently in OVN we create gateways in minions only. We can also
create a gateway in master, but we do need 2 NICs there.

This commit lets OVN handle master-only deployment. This is done
by creating a "localnet OVN gateway" on master. A localnet OVN gateway
is just a regular OVN gateway, but with a local physical interface
connected to it. When ovnkube is run with '-init-gateways -localnet-gateway',
the code will create a OVS bridge called "br-localnet" and create a internal
port called "br-nexthop". It will assign a local ip "169.254.33.2"
to OVN gateway in logical space and "169.254.33.1" to "br-localnet-nexthop"
in physical space. Now, if /proc/sys/net/ipv4/ip_forward is set to 1,
we can access the host IP. In addition, for the pods to be able to
access outside internet, a iptables NAT rule is added.

[Some of the code is copied from:
https://github.com/openvswitch/ovn-kubernetes/pull/184]
Signed-off-by: Gurucharan Shetty <guru@ovn.org>