config: Allow multi-master for clients.
The goal is to have external-ids:ovn-remote to have
"ssl:192.168.1.2:6641,ssl:192.168.1.3:6641,ssl:192.168.1.4:6641".

When running ovnkube, we will pass the address of the form
"ssl://192.168.1.2:6641,ssl://192.168.1.3:6641" to -nb-address.
The config.go takes that address and stores it in 2 OvnDBAuth
structures. One for NB DB and one for SB DB. Irrespective of
where you run ovnkube, each of this OvnDBAuth has
information for a client to use the address or for server
to listen on that address.

This commit will use the first address in the provided -nb-address
(or -sb-address) for server to listen on. And all addresses for
the client to connect to.

Signed-off-by: Gurucharan Shetty <guru@ovn.org>