policy: Use port groups.
With OVS 2.10, we have port groups. Using port groups
to do kubernetes network policy is much more simpler.
There are 3 large advantages.

1. We don't need to add ACLs to every logical port.
Instead, we add ACL once to a port group based on
a network policy. And then whenever new pods get
created, they just get added to the port group.

2. When namespace selector selects new namespaces in
a network policy, we don't have to go and apply the
change in ACL match to gazillion pods. Instead, we
just modify the ACL in the port group.

3. When a network policy is deleted, we don't need
to find all the ACLs added for that policy. Instead,
we can just delete the port group.

Signed-off-by: Gurucharan Shetty <guru@ovn.org>