WORKAROUND for network loss problem
Ovn adds the HW nic to the ovn bridge. This works as long as
ovn is running.  When docker is restarted or the ovs daemonset
is deleted the traffic continues to go to the bridge which
results in the loss of all network access. The console can
be used to repair the configuration and restore networking.

This change is a workaround that avoids placing the nic in the
bridge.  A side effect is that iptables are used.

Also, a bind mount can't be deleted in a container.
Changed the go_controller to not try. Deleted files in the
daemonset yaml files.

/var/run needs to be mounted.

ovnkube.sh now displays the git branch and commit number
when available.

Signed-off-by: Phil Cameron <pcameron@redhat.com>