cniserver: Fix HNS endpoints leak on Windows
Currently, if attaching HNS endpoint to the container fails,
the HNS endpoint does not get deleted. This happens because
the delete function receives the ID of the container instead
of the name of the HNS endpoint.

This also fixes another possible HNS endpoints leak that can
happen if running the ovs command fails.

Signed-off-by: Alin Balutoiu <alinbalutoiu@gmail.com>