config: Modify inactivity-probe for DB server and client.
OVN DB has a connection table where we currently define the
port number to listen on for the OVN DB servers.  By default,
every 5 seconds or so, OVN DB server sends a inactivity probe
to test whether the client is still connected.  When there
are hundreds of clients, the server may spend a lot of time
maintaining these inactivity probes and as a result consume
a lot of CPU. It does not make sense for the server to do
this job. It should be the client that tries to maintain
the connection. And if the connection is down, the client
should restart it.

So this commit disables the inactivity_probe for the server
by setting it as zero.

For the clients (ovn-controller in our case), we provide the
option for the user to set it. Shorter time would mean that
a lot of inactivity probes could be sent by al the clients
to the server. A very long time means that events may
not get propogated to the clients soon enough.  This commit
sets the default value as 100 seconds. And a user can change it
if he prefers.

Signed-off-by: Gurucharan Shetty <guru@ovn.org>