remove unnecessary go-routines around policy watches
With the addition of federated watch handlers, there are unnecessary go routines in the policy flow.
All we do with spinning a new go-routine for watching local peer pods for policy is 'add' the handlers and then wait for network policy to be removed so that we can 'remove' the handlers. We don't need go-routines for that.
This will be immensely helpful when we have 20k namespaces isolated from each other by at least as many NetworkPolicy objects.

Signed-off-by: Jacob Tanenbaum <jtanenba@redhat.com>