ansible: Add minion hostnames in master hosts file
The ansible playbook does not add minion hostnames inside
the /etc/hosts file on the master. As a result, the kubectl exec
command does not work since the master node cannot resolve the
minion hostname to connect to the pod.

This patch fixes this by adding the hostnames of the minion
in the master /etc/hosts file.

Signed-off-by: Alin Balutoiu <alinbalutoiu@gmail.com>