factory: Serialize callbacks for k8s events.
For integration with OVSDB golang bindings, it
makes sense to not have parallel threads trying
to access the database at the same time.

This is an attempt to serialize all events coming
from kubernetes.  This also lets us get rid of
locks and any other race conditions that we
may have in code.

Signed-off-by: Gurucharan Shetty <guru@ovn.org>