contrib: Generate automatically test yamls for Windows
The ansible playbook generates a template yaml which has to
be edited before it can be used to create windows containers.

This patch will generate automatically a yaml file for each
Windows minion in the cluster, with according tags based on OS.
This is very useful for testing out different versions of Windows.

Signed-off-by: Alin-Gheorghe Balutoiu <alinbalutoiu@gmail.com>