contrib: Create HNS network via powershell cmdlets
The HNS network gets created at the moment via docker with docker
network create command. In the future it will be possible to use
containers without Docker (either `containerd` or `CRI-O`)
We will create the HNS network via New-HnsNetwork command bundled
in the installer, to provision the network.

During minion init, a script is created which will run at startup
and Windows gets rebooted.
Since ovnkube runs as a service, this is not required anymore.

This commit removes the Windows reboot and starts directly the
ovn-kubernetes-node service which will run minion init. The restart
after docker installation is also removed since the path is
already updated.

Signed-off-by: Alin-Gheorghe Balutoiu <alinbalutoiu@gmail.com>
Signed-off-by: Alin Gabriel Serdean <aserdean@ovn.org>
Co-authored-by: Alin Gabriel Serdean <aserdean@ovn.org>