contrib: Run the startup script via psexec
The Run registry key doesn't exist if user didn't login
into the system. This causes the playbook to fail when
it is trying to register the script to be executed at
startup.

To address this issue, the script is now being executed
with psexec. This also allows to remove the reboot which
is not required anymore.

Signed-off-by: Alin-Gheorghe Balutoiu <alinbalutoiu@gmail.com>