Fixed race condition with the informer handlers
When trying to implement network policy for Namespace and Pod selectors a race
condition was discovered with the informer lock and namespace policy locks.
Specifically this manifested when the network policy delete grabbed the namespace
policy lock and deadlocked waiting for the pod informer lock from an ongoing pod delete.

Copying the informer handlers, releaseing the lock, and calling the copies to
do the work solves this problem.

Signed-off-by: Jacob Tanenbaum <jtanenba@redhat.com>