factory: defer handler removal to a goroutine
It's possible to deadlock if two events come in quick succession,
and one of the events triggers a watch handler removal. In one
specific case:

1) NetworkPolicy added -> addNetworkPolicyOld() -> handleLocalPodSelectorOld()
    ∘ adds pod event handler
2) Pod is deleted from API
    ∘ Pod informer Lock() grabbed
    ∘ handleLocalPodSelectorDelFuncOld()
        ∘ tries to grab np.Lock()
3) NetworkPolicy deleted -> deleteNetworkPolicyOld()
    ∘ grabs np.Lock()
    ∘ shutdownHandlers()
       ∘ RemovePodHandler()
           ‣ tries to grab Pod informer Lock()

Fix the deadlock by setting a tombstone on handlers so they
are not called after they are "removed", but do the removal
from a separate goroutine that can lock/unlock the informer
independently of any of the actual OVN logic.

Signed-off-by: Dan Williams <dcbw@redhat.com>